setup enemies, recreating world map
