+++ src/App.js
@@ -3,11 +3,7 @@ import World from "./features/world";
 
 class App extends Component {
   render() {
-    return (
-      <div>
-        <World />
-      </div>
-    );
+    return <World />;
   }
 }
