+++ src/maps/1/enemies.js
@@ -1,0 +1,22 @@
+export const enemies = [
+  {
+    name: "Wolf",
+    image: "Wolf.png",
+    str: 16,
+    int: 9,
+    wisdom: 9,
+    dex: 12,
+    const: 18
+  },
+  {
+    name: "Spider",
+    image: "spider.png",
+    str: 10,
+    int: 15,
+    wisdom: 13,
+    dex: 14,
+    const: 11
+  }
+];
+
+export default enemies;
