+++ src/maps/1/index.js
@@ -1,15 +1,10 @@
-// prettier-ignore
-export const tiles = [
-    [0,0,0,5,5,5,5,5,6,6,6,6,6,6,6,6,6,6,6,6],
-    [6,0,0,5,5,5,5,6,6,6,6,6,6,6,6,6,6,5,5,6],
-    [6,6,0,0,6,5,6,6,0,0,6,6,0,0,6,6,0,4,5,6],
-    [6,6,0,0,6,6,6,6,6,0,0,0,0,5,6,6,0,0,5,6],
-    [6,6,0,0,0,6,6,6,6,0,0,6,5,5,6,6,6,3,6,6],
-    [6,5,5,5,0,0,6,0,6,0,0,6,5,6,6,6,6,3,6,6],
-    [6,6,5,5,0,0,0,0,0,0,0,6,6,6,0,0,0,0,6,6],
-    [6,6,6,6,6,0,0,6,6,6,0,0,6,0,0,6,6,0,6,5],
-    [6,6,6,0,0,0,0,6,6,0,0,0,0,0,0,5,6,6,5,5],
-    [6,6,6,6,0,0,6,6,6,0,0,6,0,0,5,5,6,6,5,5],
-    [6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,5,5],
-    [6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,5,5,5]
-]
+import tiles from "./tiles";
+import enemies from "./enemies";
+
+const levelRange = [1, 4];
+
+export default {
+  tiles,
+  enemies,
+  levelRange
+};
