+++ src/maps/1/tiles.js
@@ -1,0 +1,15 @@
+// prettier-ignore
+export const tiles = [
+    [0,0,0,5,5,5,5,5,6,6,6,6,6,6,6,6,6,6,6,6],
+    [6,0,0,5,5,5,5,6,6,6,6,6,6,6,6,6,6,5,5,6],
+    [6,6,0,0,6,5,6,6,0,0,6,6,0,0,6,6,0,4,5,6],
+    [6,6,0,0,6,6,6,6,6,0,0,0,0,5,6,6,0,0,5,6],
+    [6,6,0,0,0,6,6,6,6,0,0,6,5,5,6,6,6,3,6,6],
+    [6,5,5,5,0,0,6,0,6,0,0,6,5,6,6,6,6,3,6,6],
+    [6,6,5,5,0,0,0,0,0,0,0,6,6,6,0,0,0,0,6,6],
+    [6,6,6,6,6,0,0,6,6,6,0,0,6,0,0,6,6,0,6,5],
+    [6,6,6,0,0,0,0,6,6,0,0,0,0,0,0,5,6,6,5,5],
+    [6,6,6,6,0,0,6,6,6,0,0,6,0,0,5,5,6,6,5,5],
+    [6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,5,5],
+    [6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,5,5,5]
+]
